Lots of complicated text-wrapping logic, but now we word-wrap and add thumbnail images to the output
Adds the ability to download thumbnails, and a whole lot of word-wrapping logic (not perfect for sure - need to know if requires a static height+width - could try downsizing the fonts if that'd work better). But now we use bing images to find a thumbnail, save it, and then we add it on top of the text.

TODO: check if the thumbnail image actually exists