Convert files into classes, integrate pyforms, start basic GUI
Integrated pyforms-gui and created a mainApplication class that creates a couple buttons, a text field and a file picker. You can now generate a spreadsheet by entering a list of words (one per line) and you can select a spreadsheet and create output images from it. No integration with the thumbnails yet.

- Updated the readme to include pyforms